backend/requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.16
django-colorfield==0.9.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.7.0
executing==1.2.0
filetype==1.2.0
idna==3.4
ipython==8.14.0
itypes==1.2.0
jedi==0.19.0
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.6
